Convert getHoldHarmlessPreamble from function to component
