+++ src/Pages/Volunteer/components/GetHoldHarmlessPreamble/index.js
@@ -1,0 +1,1 @@
+export { GetHoldHarmlessPreamble } from "./GetHoldHarmlessPreamble";
